+++ changelog.md
@@ -7,6 +7,16 @@ This project follows semVer, where:
 <!-- ### [unreleased]
 -->
 
+### v6.16.3 [Aug 2021]
+
+- **[new]** - .daysInMonth() method (thanks m1212e!)
+- **[change]** - replace deprecated timezone aliases (thanks Michael!)
+- **[change]** - date/day naming inconsistency in add method (thanks Andy!)
+- **[fix]** - typescript fixes (thanks Adam!)
+- **[fix]** - many-year add issue #235
+- **[fix]** - iso format year-padding for BC years
+- update deps
+
 ### v6.16.2 [June 2021]
 
 - **[fix]** - offset for `asia/baku` #292